+++ frontend/src/components/AddTask.tsx
@@ -1,12 +1,12 @@
 import { useState, useEffect } from "react";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardFooter } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+import { Button } from "@/components/ui/Button";
+import { Card, CardContent, CardFooter } from "@/components/ui/Card";
+import { Input } from "@/components/ui/Input";
+import { Label } from "@/components/ui/Label";
 import { toast } from "sonner";
 import axios from "axios";
 import useStore from "@/utils/store";
-import { DatePickerDemo } from "./ui/datepicker";
+import { DatePickerDemo } from "./ui/Datepicker";
 
 interface data {
   title: string;
