+++ frontend/src/components/ui/calender.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { DayPicker } from "react-day-picker";
 
 import { cn } from "@/lib/utils";
-import { buttonVariants } from "@/components/ui/button";
+import { buttonVariants } from "@/components/ui/Button";
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
