+++ frontend/src/pages/auth/Login.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Button } from "@/components/ui/Button";
+import { Input } from "@/components/ui/Input";
+import { Label } from "@/components/ui/Label";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/Card";
 import { useNavigate } from "react-router-dom";
 import { Toaster, toast } from "sonner";
 import axios from "axios";
@@ -69,7 +69,7 @@ export default function Login() {
         email: response.data.data.email,
       });
       toast.success("Login successful!");
-      navigate("/profile");
+      navigate("/home");
     } catch (error: unknown) {
       console.error(error);
       toast.error("Invalid credentials");
